Added interface for Customer
